feat: modularize server with routes, services, and config structure
